requirements.txt
# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Deep Learning and Image Processing
torch==2.1.0
torchvision==0.16.0
pillow==10.0.1

# Core scientific libraries
numpy==1.24.3
pandas==2.0.3

# Development and debugging
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Optional: for enhanced logging and monitoring
loguru==0.7.2
